Created CDISC fixture for initial project state needed in script 06 - Assign User Rights
Project has most of the requirements in the manual script, reducing setup time. It is not a perfect copy of what should be the end result of the previous script (which is what the script calls for). One issue with it currently is that the record with a survey response has its survey form status as "Complete" instead of "Survey  Response Complete" which causes one of the scenarios to fail. See comment for scenario 26 in the feature file for more context.